+++ Octokit.Reactive/Clients/IObservableNotificationsClient.cs
@@ -40,7 +40,7 @@ namespace Octokit.Reactive
         /// </summary>
         /// <param name="repositoryId">The Id of the repository.</param>
         /// <exception cref="AuthorizationException">Thrown if the client is not authenticated.</exception>
-        IObservable<Notification> GetAllForRepository(int repositoryId);
+        IObservable<Notification> GetAllForRepository(long repositoryId);
 
         /// <summary>
         /// Retrieves all of the <see cref="Notification"/>s for the current user specific to the specified repository.
@@ -57,7 +57,7 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository.</param>
         /// <param name="options">Options for changing the API response</param>
         /// <exception cref="AuthorizationException">Thrown if the client is not authenticated.</exception>
-        IObservable<Notification> GetAllForRepository(int repositoryId, ApiOptions options);
+        IObservable<Notification> GetAllForRepository(long repositoryId, ApiOptions options);
 
         /// <summary>
         /// Retrieves all of the <see cref="Notification"/>s for the current user.
@@ -91,7 +91,7 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository.</param>
         /// <param name="request">Specifies the parameters to filter notifications by</param>
         /// <exception cref="AuthorizationException">Thrown if the client is not authenticated.</exception>
-        IObservable<Notification> GetAllForRepository(int repositoryId, NotificationsRequest request);
+        IObservable<Notification> GetAllForRepository(long repositoryId, NotificationsRequest request);
 
         /// <summary>
         /// Retrieves all of the <see cref="Notification"/>s for the current user specific to the specified repository.
@@ -110,7 +110,7 @@ namespace Octokit.Reactive
         /// <param name="request">Specifies the parameters to filter notifications by</param>
         /// <param name="options">Options for changing the API response</param>
         /// <exception cref="AuthorizationException">Thrown if the client is not authenticated.</exception>
-        IObservable<Notification> GetAllForRepository(int repositoryId, NotificationsRequest request, ApiOptions options);
+        IObservable<Notification> GetAllForRepository(long repositoryId, NotificationsRequest request, ApiOptions options);
 
         /// <summary>
         /// Marks all notifications as read.
@@ -138,7 +138,7 @@ namespace Octokit.Reactive
         /// </summary>
         /// <param name="repositoryId">The Id of the repository</param>
         /// <remarks>http://developer.github.com/v3/activity/notifications/#mark-notifications-as-read-in-a-repository</remarks>
-        IObservable<Unit> MarkAsReadForRepository(int repositoryId);
+        IObservable<Unit> MarkAsReadForRepository(long repositoryId);
 
         /// <summary>
         /// Marks the notifications for a given repository as read.
@@ -155,7 +155,7 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="markAsReadRequest">The <see cref="MarkAsReadRequest"/> parameter which specifies which notifications to mark.</param>
         /// <remarks>http://developer.github.com/v3/activity/notifications/#mark-notifications-as-read-in-a-repository</remarks>
-        IObservable<Unit> MarkAsReadForRepository(int repositoryId, MarkAsReadRequest markAsReadRequest);
+        IObservable<Unit> MarkAsReadForRepository(long repositoryId, MarkAsReadRequest markAsReadRequest);
 
         /// <summary>
         /// Retrives a single <see cref="Notification"/> by Id.
